+++ pkg/manager/volumes/pod_vol_modifier.go
@@ -11,6 +11,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	errutil "k8s.io/apimachinery/pkg/util/errors"
+	storagelister "k8s.io/client-go/listers/storage/v1"
 	klog "k8s.io/klog/v2"
 
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
@@ -20,6 +21,9 @@ import (
 )
 
 type PodVolumeModifier interface {
+	GetDesiredVolumes(tc *v1alpha1.TidbCluster, mt v1alpha1.MemberType) ([]DesiredVolume, error)
+	GetActualVolumes(pod *corev1.Pod, vs []DesiredVolume) ([]ActualVolume, error)
+
 	Modify(tc *v1alpha1.TidbCluster, pod *corev1.Pod, expected []DesiredVolume, shouldEvictLeader bool) (bool, error)
 }
 
@@ -54,7 +58,7 @@ func NewPodVolumeModifier(deps *controller.Dependencies) PodVolumeModifier {
 func (p *podVolModifier) Modify(tc *v1alpha1.TidbCluster, pod *corev1.Pod, expected []DesiredVolume, shouldEvictLeader bool) (bool, error) {
 	ctx := context.TODO()
 
-	actual, err := p.getActualVolumes(pod, expected)
+	actual, err := p.GetActualVolumes(pod, expected)
 	if err != nil {
 		return false, err
 	}
@@ -117,6 +121,114 @@ func (p *podVolModifier) Modify(tc *v1alpha1.TidbCluster, pod *corev1.Pod, expec
 	return waitForLeaderEviction, errutil.NewAggregate(errs)
 }
 
+// TODO: it should be refactored
+func (p *podVolModifier) GetDesiredVolumes(tc *v1alpha1.TidbCluster, mt v1alpha1.MemberType) ([]DesiredVolume, error) {
+	desiredVolumes := []DesiredVolume{}
+	scLister := p.deps.StorageClassLister
+
+	storageVolumes := []v1alpha1.StorageVolume{}
+	switch mt {
+	case v1alpha1.PDMemberType:
+		sc, err := getStorageClass(tc.Spec.PD.StorageClassName, scLister)
+		if err != nil {
+			return nil, err
+		}
+		name := v1alpha1.GetStorageVolumeName("", mt)
+		size := getStorageSize(tc.Spec.PD.Requests)
+		d := DesiredVolume{
+			Name:         string(name),
+			Size:         size.String(),
+			StorageClass: sc,
+		}
+		desiredVolumes = append(desiredVolumes, d)
+
+		storageVolumes = tc.Spec.PD.StorageVolumes
+
+	case v1alpha1.TiDBMemberType:
+		storageVolumes = tc.Spec.TiDB.StorageVolumes
+
+	case v1alpha1.TiKVMemberType:
+		sc, err := getStorageClass(tc.Spec.TiKV.StorageClassName, scLister)
+		if err != nil {
+			return nil, err
+		}
+		name := v1alpha1.GetStorageVolumeName("", mt)
+		size := getStorageSize(tc.Spec.TiKV.Requests)
+		d := DesiredVolume{
+			Name:         string(name),
+			Size:         size.String(),
+			StorageClass: sc,
+		}
+		desiredVolumes = append(desiredVolumes, d)
+
+		storageVolumes = tc.Spec.TiKV.StorageVolumes
+
+	case v1alpha1.TiFlashMemberType:
+		for i, claim := range tc.Spec.TiFlash.StorageClaims {
+			sc, err := getStorageClass(claim.StorageClassName, scLister)
+			if err != nil {
+				return nil, err
+			}
+			name := v1alpha1.GetStorageVolumeNameForTiFlash(i)
+			size := getStorageSize(claim.Resources.Requests)
+			d := DesiredVolume{
+				Name:         string(name),
+				Size:         size.String(),
+				StorageClass: sc,
+			}
+			desiredVolumes = append(desiredVolumes, d)
+		}
+
+	case v1alpha1.TiCDCMemberType:
+		storageVolumes = tc.Spec.TiCDC.StorageVolumes
+
+	case v1alpha1.PumpMemberType:
+		sc, err := getStorageClass(tc.Spec.Pump.StorageClassName, scLister)
+		if err != nil {
+			return nil, err
+		}
+		name := v1alpha1.GetStorageVolumeName("", mt)
+		size := getStorageSize(tc.Spec.Pump.Requests)
+		d := DesiredVolume{
+			Name:         string(name),
+			Size:         size.String(),
+			StorageClass: sc,
+		}
+		desiredVolumes = append(desiredVolumes, d)
+	default:
+		return nil, fmt.Errorf("unsupported member type %s", mt)
+	}
+
+	for _, sv := range storageVolumes {
+		if quantity, err := resource.ParseQuantity(sv.StorageSize); err == nil {
+			sc, err := getStorageClass(sv.StorageClassName, scLister)
+			if err != nil {
+				return nil, err
+			}
+			name := v1alpha1.GetStorageVolumeName(sv.Name, mt)
+			d := DesiredVolume{
+				Name:         string(name),
+				Size:         quantity.String(),
+				StorageClass: sc,
+			}
+
+			desiredVolumes = append(desiredVolumes, d)
+
+		} else {
+			klog.Warningf("StorageVolume %q in %s .spec.%s is invalid", sv.Name, getTcKey(tc), mt)
+		}
+	}
+
+	return desiredVolumes, nil
+}
+
+func getStorageClass(name *string, scLister storagelister.StorageClassLister) (*storagev1.StorageClass, error) {
+	if name == nil {
+		return nil, nil
+	}
+	return scLister.Get(*name)
+}
+
 func getDesiredVolumeByName(vs []DesiredVolume, name string) *DesiredVolume {
 	for i := range vs {
 		v := &vs[i]
@@ -146,7 +258,7 @@ func (p *podVolModifier) getPVC(ns string, vol *corev1.Volume) (*corev1.Persiste
 	return pvc, nil
 }
 
-func (p *podVolModifier) getActualVolumes(pod *corev1.Pod, vs []DesiredVolume) ([]ActualVolume, error) {
+func (p *podVolModifier) GetActualVolumes(pod *corev1.Pod, vs []DesiredVolume) ([]ActualVolume, error) {
 	vols := []ActualVolume{}
 
 	for i := range pod.Spec.Volumes {
